feat: Saving data from form to Google Spreadsheet
It is possible to write scratched data to Google Spreadsheet.
Every scrached item is added to the defined sheet.
Sheet's id and tab name can be defined in plugin main form.
If user is not signed in to Google account - login form is provided by Google, as well as consent form
for connecting with user's data.