refactor: validate form sections independently
Form parts can be validated separately
- fields from facebook and fields for selected sending option
  (e.g. Google Sheets or Cooper).
  The reason was to separate validating selected option data
 from deselected ones.
Function isTheFormCorrect was renamed to isTheFormIncorrect
because of readability (returns fales if no errors).
isTheFormIncorrect() now recives formItem
(which is form or fieldset element) to be validated
hence can be used to validating different parts of form.
form#scratch-me-form is changed to use fieldsets for independent validation